+++ src/extension.ts
@@ -109,6 +109,9 @@ export function activate(context: vscode.ExtensionContext): void {
     vscode.commands.registerCommand("rojoExplorer.createModel", (node?: ExplorerNode) =>
       createResource(provider, fileSystem, node, "Model"),
     ),
+    vscode.commands.registerCommand("rojoExplorer.createRemoteEvent", (node?: ExplorerNode) =>
+      createResource(provider, fileSystem, node, "RemoteEvent"),
+    ),
     vscode.commands.registerCommand("rojoExplorer.createStringValue", (node?: ExplorerNode) =>
       createResource(provider, fileSystem, node, "StringValue"),
     ),
@@ -775,6 +778,8 @@ function localizeResourceKind(kind: CreatableResourceKind): string {
       return vscode.l10n.t("ModuleScript");
     case "Model":
       return vscode.l10n.t("Model");
+    case "RemoteEvent":
+      return vscode.l10n.t("RemoteEvent");
     case "StringValue":
       return vscode.l10n.t("StringValue");
     case "LocalizationTable":